Se termina la arquitectura global y se empieza a definir el draw item

+++ coloUi/colouibase.cpp
@@ -1,7 +1,7 @@
 #include "colouibase.h"
 
-qreal ColoUiBase::SCREEN_WIDTH = 0;
-qreal ColoUiBase::SCREEN_HEIGHT = 0;
+quint16 ColoUiBase::SCREEN_WIDTH = 0;
+quint16 ColoUiBase::SCREEN_HEIGHT = 0;
 
 ColoUiBase::ColoUiBase()
 {
